I have added a readme.txt which is pretty basic but gives you an idea.
I have also tried to make the javascript file ES6 compliant.  This has 
been largly achieved so far by changing the variables from var to let or 
const where appropriate - going to have to look for a transpiler for 
backwards compatability.

Finally, I have used boostrap 4 in the html file so that its fairly 
responsive to differnt browser sizes - a tad overkill for a simple page, 
but it might be useful to drop some of the code into a page for someone 
in future.